+++ processor/src/test/resources/input/com/example/component_dependency/RuntimeTypeValidateDependency.java
@@ -1,0 +1,14 @@
+package com.example.component_dependency;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentDependency;
+
+@ArezComponent
+public abstract class RuntimeTypeValidateDependency
+{
+  @ComponentDependency( validateTypeAtRuntime = true )
+  public final Object getValue()
+  {
+    return null;
+  }
+}